requirements.txt
b2sdk==1.21.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
dnspython==2.3.0
google-api-core==2.11.0
google-auth==2.19.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
idna==3.4
logfury==1.0.1
mongoengine==0.27.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pymongo==4.3.3
requests==2.31.0
rsa==4.9
six==1.16.0
tqdm==4.65.0
urllib3==1.26.16